@@ -13,46 +13,46 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "^7.2.2",
-    "@angular/core": "^7.2.2",
-    "@angular/forms": "^7.2.2",
-    "@angular/http": "^7.2.2",
-    "@angular/platform-browser": "^7.2.2",
-    "@angular/platform-browser-dynamic": "^7.2.2",
-    "@angular/router": "^7.2.2",
-    "@ionic-native/core": "^5.0.0",
-    "@ionic-native/splash-screen": "^5.0.0",
-    "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^5.3.4",
-    "core-js": "^2.5.4",
-    "rxjs": "~6.6.3",
-    "zone.js": "~0.8.29"
+    "@angular/common": "^10.2.4",
+    "@angular/core": "^10.2.4",
+    "@angular/forms": "^10.2.4",
+    "@angular/http": "^7.2.16",
+    "@angular/platform-browser": "^10.2.4",
+    "@angular/platform-browser-dynamic": "^10.2.4",
+    "@angular/router": "^10.2.4",
+    "@ionic-native/core": "^5.31.1",
+    "@ionic-native/splash-screen": "^5.31.1",
+    "@ionic-native/status-bar": "^5.31.1",
+    "@ionic/angular": "^5.6.0",
+    "core-js": "^2.6.12",
+    "rxjs": "^6.6.6",
+    "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/architect": "^0.1001.3",
-    "@angular-devkit/build-angular": "^0.1001.4",
-    "@angular-devkit/core": "^10.1.4",
-    "@angular-devkit/schematics": "^10.1.4",
-    "@angular/cli": "^10.1.4",
-    "@angular/compiler": "~7.2.2",
-    "@angular/compiler-cli": "^10.1.3",
-    "@angular/language-service": "~7.2.2",
+    "@angular-devkit/architect": "^0.1001.7",
+    "@angular-devkit/build-angular": "^0.1102.3",
+    "@angular-devkit/core": "^10.2.3",
+    "@angular-devkit/schematics": "^10.2.3",
+    "@angular/cli": "^10.2.3",
+    "@angular/compiler": "^10.1.6",
+    "@angular/compiler-cli": "^10.2.4",
+    "@angular/language-service": "^10.1.6",
     "@ionic/angular-toolkit": "^2.3.3",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~10.12.0",
+    "@types/jasmine": "^2.8.17",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^10.12.30",
     "codelyzer": "~4.5.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "^5.2.3",
     "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-coverage-istanbul-reporter": "^2.0.6",
     "karma-jasmine": "~1.1.2",
     "karma-jasmine-html-reporter": "^0.2.2",
     "protractor": "^7.0.0",
-    "ts-node": "~8.0.0",
+    "ts-node": "^8.0.3",
     "tslint": "~5.12.0",
-    "typescript": "~4.0.3"
+    "typescript": "^4.0.7"
   },
   "description": "An Ionic project"
 }
